Remove runtime dependency
